# replace aho corasick 