--- a/go.mod
+++ b/go.mod
@@ -3,9 +3,9 @@ module github.com/corazawaf/coraza/v2
 go 1.16
 
 require (
-	github.com/cloudflare/ahocorasick v0.0.0-20210425175752-730270c3e184
 	github.com/corazawaf/libinjection-go v0.0.0-20220207031228-44e9c4250eb5
 	github.com/foxcpp/go-mockdns v1.0.0
+	github.com/petar-dambovaliev/aho-corasick v0.0.0-20211021192214-5ab2d9280aa9
 	go.uber.org/atomic v1.9.0 // indirect
 	go.uber.org/multierr v1.8.0 // indirect
 	go.uber.org/zap v1.21.0