This is a huge improvement, memory consumption is 60% less, and execution time was 233% faster

## Benchmark results using CRS:

**Old results:**
```
Preparing CRS...
CRS PATH: /var/folders/tf/23czq0cd4wd5v54zcvctvtzw0000gn/T/crs1564927737
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/testing
BenchmarkCRSCompilation-10             3         378610667 ns/op
BenchmarkCRSSimpleGET-10             613           1791376 ns/op
BenchmarkCRSSimplePOST-10            308           3499975 ns/op
PASS
ok      github.com/corazawaf/coraza/v3/testing  13.985s
```

**New results:**
```
Preparing CRS...
CRS PATH: /var/folders/tf/23czq0cd4wd5v54zcvctvtzw0000gn/T/crs1721984072
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/testing
BenchmarkCRSCompilation-10            21          48269002 ns/op
BenchmarkCRSSimpleGET-10             777           1396721 ns/op
BenchmarkCRSSimplePOST-10            489           2396956 ns/op
PASS
ok      github.com/corazawaf/coraza/v3/testing  5.969s
```