@jptosso 2022-08-19
> @anuraaga worth for you to have a look at github.com/petar-dambovaliev/aho-corasick repository to see if all effort towards performance is enough?

Looks quite similar to our implementation, we would have to run benchmarks.